Refactor logger and version handling
Refactored the logger package to use a new configuration structure, improved log level handling, and updated internal logging logic. Replaced string-based version handling with a types.Version struct throughout the codebase, including logger, configs, and controller packages. Updated all deprecation and logging calls to use the new version type and logger API. Removed unused code and improved error handling and formatting in the logger.

+++ src/utils/file.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 )
 
-func CallOrigin(depth int) (string, int) {
+func CallOrigin(depth int) (string, uint64) {
 	_, file, line, ok := runtime.Caller(depth)
 
 	if ok {
@@ -16,5 +16,5 @@ func CallOrigin(depth int) (string, int) {
 		line = 0
 	}
 
-	return file, line
+	return file, uint64(line)
 }
